In pal2ical, skip comments then parse event
Since the last thing that happens in the file stream is an event
gets parsed, we need to skip comments to get to the next event
to be parsed.  This is the standard order used in the main program
and the only way that works correctly if the line after the first
event is a comment.